web/backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
flask-cors==4.0.0

# Computer Vision and ML
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
scikit-learn==1.2.0
lightgbm==4.6.0

# Image processing
Pillow==10.1.0

# Environment and utilities
python-dotenv==1.0.0
requests==2.31.0

# Redis for session management and caching
redis==5.0.1

# Async worker support (opsiyonel - kurulum sorunları varsa yoruma alın)
# gevent==24.2.1

# Hugging Face Spaces için ek optimizasyonlar
gunicorn==21.2.0